refactor: move scalar math to primitives (MOR-2382)
