+++ frontend/fixtures/approved-baselines/README.md
@@ -29,8 +29,8 @@ Tracked, reviewed screenshots compared by `npm run test:e2e:visual` (`playwright
 | `gallery-statusindicator` | MOR-2219 looks gallery — preserves StatusIndicator's look, carried by the `.v2-status-indicator` rule set in `control-button.css` |
 | `gallery-valuecontrol-hbar` | MOR-2219 looks gallery — preserves the HBar renderer's look; carried by `components-v2/controls/value-control/HBarRenderer.svelte` (owns its own track/thumb div geometry — no SVG) |
 | `gallery-valuecontrol-bipolar` | MOR-2219 looks gallery — preserves the Bipolar renderer's look; carried by `BipolarRenderer.svelte` |
-| `gallery-valuecontrol-knob` | MOR-2219 looks gallery — preserves the Knob renderer's look; carried by `KnobRenderer.svelte` (builds its own `<svg>` arc geometry via `value-control-core.ts` helpers) |
-| `gallery-professionalknob` | MOR-2219 looks gallery — preserves the ProfessionalKnob skin's look; carried by `value-control/skins/ProfessionalKnob.svelte` itself — its arc/indicator/tick geometry comes from the same shared `value-control-core.ts` helpers (`calculateArcPath`, `calculateIndicatorPosition`, `generateTickPositions`) that `KnobRenderer.svelte` uses, but its SVG markup (radial gradients, glow filter, element structure) and `.pro-*` styles are local to this file, not shared with any other renderer or skin |
+| `gallery-valuecontrol-knob` | MOR-2219 looks gallery — preserves the Knob renderer's look; carried by `KnobRenderer.svelte` (builds its own `<svg>` arc geometry via `primitives/scalar/value-control-core.ts` helpers) |
+| `gallery-professionalknob` | MOR-2219 looks gallery — preserves the ProfessionalKnob skin's look; carried by `value-control/skins/ProfessionalKnob.svelte` itself — its arc/indicator/tick geometry comes from the same shared `primitives/scalar/value-control-core.ts` helpers (`calculateArcPath`, `calculateIndicatorPosition`, `generateTickPositions`) that `KnobRenderer.svelte` uses, but its SVG markup (radial gradients, glow filter, element structure) and `.pro-*` styles are local to this file, not shared with any other renderer or skin |
 | `peer-split-chassis--desktop` | MOR-2243 — the segmentline peer-split glass chassis (`PeerSplitLayout.svelte`, the only production file that mounts `ScaledStage`); at this frame size its stage scale resolves to exactly 1, the identity transform |
 | `peer-split-chassis--1100x800` | MOR-2243 — the same chassis at the only frame size in this set where `computeStageScale` resolves below 1 (1100/1280 ≈ 0.859), so the stage transform is exercised rather than skipped |
 | `unified-instrument--desktop` | MOR-2325 B/D — segmentline LCD unified-instrument composition at the standard 1280×800 desktop viewport |
