+++ frontend/src/components-v2/controls/value-control/BipolarRenderer.svelte
@@ -12,7 +12,7 @@
     clamp,
     snapToStep,
     valueToPosition,
-  } from './value-control-core';
+  } from '../../../primitives/scalar/value-control-core';
 
   interface Props {
     value: number;
