+++ frontend/src/components-v2/controls/value-control/HBarRenderer.svelte
@@ -10,7 +10,7 @@
     debounce,
     clamp,
     snapToStep,
-  } from './value-control-core';
+  } from '../../../primitives/scalar/value-control-core';
 
   interface Props {
     value: number;
