+++ frontend/src/components-v2/controls/value-control/index.ts
@@ -4,6 +4,5 @@ export { default as BipolarRenderer } from './BipolarRenderer.svelte';
 export { default as DualParamRenderer } from './DualParamRenderer.svelte';
 export { default as KnobRenderer } from './KnobRenderer.svelte';
 export { default as DiscreteRenderer } from './DiscreteRenderer.svelte';
-export * from './value-control-core';
 export type { Skin, SkinRendererProps, KnobSkinRendererProps } from './skin';
 export { professionalSkin, skins } from './skins/index';
