+++ frontend/src/components-v2/panels/VoxPanel.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-  import { ValueControl, rawToPercentDisplay } from '../controls/value-control';
+  import { ValueControl } from '../controls/value-control';
+  import { rawToPercentDisplay } from '../../primitives/scalar/value-control-core';
   import { HardwareButton } from '$lib/Button';
 
   interface Props {
